backend/app/dashboard_api/serializers/app_authenticate_serializer.py
from rest_framework import serializers


class AppAuthenticateSerializer(serializers.Serializer):
    username = serializers.CharField(required=True, max_length=255)
    password = serializers.CharField(required=True, max_length=255)
